+++ CHANGELOG.md
@@ -47,6 +47,7 @@ BUG FIXES:
   * provisioner/ansible: set cwd to staging directory [GH-1016]
   * provisioners/chef-client: Don't chown directory with Ubuntu. [GH-939]
   * provisioners/shell: Env var values can have equal signs. [GH-1045]
+  * provisioners/shell: chmod the uploaded script file to 0777. [GH-994]
   * post-processor/docker-push: Allow repositories with ports. [GH-923]
   * post-processor/vagrant: Create parent directories for `output` path [GH-1059]